Creando nueva clase main para organizar las listas

+++ src/com/aluracurson/screenmatch/calculos/Calificacion.java
@@ -1,5 +1,5 @@
 package com.aluracurson.screenmatch.calculos;
 
 public interface Calificacion {
-    int getCalificacion();
+    double getCalificacion();
 }
